// src/types/game/item.d.ts
export type ItemType =
    | 'Item'
    | 'Fuel'
    | 'Other'
    | 'BuildingMaterial'
    | 'Tool'
    | 'HouseholdGoods'
    | 'Lubricant'
    | 'Battery'
    | 'Electronics'
    | 'MedicalSupplies'
    | 'Jewelry'
    | 'Pms'
    | 'FlashHider'
    | 'Silencer'
    | 'MuzzleCombo'
    | 'Compensator'
    | 'ThermalVision'
    | 'NightVision'
    | 'AssaultScope'
    | 'OpticScope'
    | 'Collimator'
    | 'CompactCollimator'
    | 'IronSight'
    | 'SpecialScope'
    | 'Foregrip'
    | 'Bipod'
    | 'AuxiliaryMod'
    | 'Gasblock'
    | 'Sights'
    | 'Flashlight'
    | 'LightLaser'
    | 'TacticalCombo'
    | 'RailCovers'
    | 'Muzzle'
    | 'CylinderMagazine'
    | 'Shaft'
    | 'Charge'
    | 'Stock'
    | 'Launcher'
    | 'Magazine'
    | 'Mount'
    | 'Barrel'
    | 'MasterMod'
    | 'PistolGrip'
    | 'Receiver'
    | 'Handguard'
    | 'FunctionalMod'
    | 'GearMod'
    | 'Pockets'
    | 'Backpack'
    | 'MobContainer'
    | 'LootContainer'
    | 'Vest'
    | 'SortingTable'
    | 'SpecialWeapon'
    | 'SniperRifle'
    | 'AssaultCarbine'
    | 'Pistol'
    | 'Shotgun'
    | 'GrenadeLauncher'
    | 'Smg'
    | 'AssaultRifle'
    | 'MachineGun'
    | 'MarksmanRifle'
    | 'Revolver'
    | 'Armor'
    | 'Visors'
    | 'FaceCover'
    | 'Headwear'
    | 'Headphones'
    | 'ArmoredEquipment'
    | 'Stash'
    | 'SearchableItem'
    | 'ArmBand'
    | 'Weapon'
    | 'LockableContainer'
    | 'SimpleContainer'
    | 'Inventory'
    | 'Equipment'
    | 'StationaryContainer'
    | 'Mod'
    | 'Drink'
    | 'Food'
    | 'Keycard'
    | 'KeyMechanical'
    | 'Stimulator'
    | 'MedKit'
    | 'Medical'
    | 'Compass'
    | 'Drugs'
    | 'PortableRangeFinder'
    | 'Money'
    | 'AmmoBox'
    | 'Ammo'
    | 'BarterItem'
    | 'CompoundItem'
    | 'SpecItem'
    | 'Knife'
    | 'ThrowWeap'
    | 'Map'
    | 'StackableItem'
    | 'Info'
    | 'Key'
    | 'FoodDrink'
    | 'Meds'

type TarkovToolsTypes =
    | 'ammo'
    | 'ammoBox'
    | 'armor'
    | 'backpack'
    | 'barter'
    | 'container'
    | 'disabled'
    | 'glasses'
    | 'grenade'
    | 'gun'
    | 'headphones'
    | 'helmet'
    | 'injectors'
    | 'keys'
    | 'markedOnly'
    | 'meds'
    | 'mods'
    | 'noFlea'
    | 'pistolGrip'
    | 'provisions'
    | 'rig'
    | 'suppressor'
    | 'wearable'

// Tarkov Tools
export interface TarkovToolsItem {
    id: string
    shortName: string
    name: string
    types: TarkovToolsTypes[]
    updated: string
    avg24hPrice: number
    lastLowPrice: number | null
    changeLast48hPercent: number | null
    width: number
    height: number
    wikiLink: string
    sellFor: ItemPrice[]
    buyFor: ItemPrice[]
}

export interface ItemPrice {
    source: TraderName
    price: number
    requirements: Requirement[]
}

export interface Requirement {
    type: RequirementType
    value: number | null
}

export interface RawItem {
    _id: string
    _name: string
    _parent: string
    _type: Type
    _props: RawItemProps
    _proto?: string
}
// ...
enum Type {
    Item = 'Item',
    Node = 'Node'
}

const enum ArmorMaterial {
    Aluminium = 'Aluminium',
    Aramid = 'Aramid',
    ArmoredSteel = 'ArmoredSteel',
    Ceramic = 'Ceramic',
    Combined = 'Combined',
    Glass = 'Glass',
    Titan = 'Titan',
    Uhmwpe = 'UHMWPE'
}

const enum TraderName {
    Prapor = 'prapor',
    Therapist = 'therapist',
    Fence = 'fence',
    Skier = 'skier',
    Peacekeeper = 'peacekeeper',
    Mechanic = 'mechanic',
    Ragman = 'ragman',
    Jaeger = 'jaeger',
    FleaMarket = 'fleaMarket'
}